♻ Split build_versions out into multiple files
⬆ Bump to python 3.9
♻ Replace requests-html with beatifulsoup4
🚨 Lint using black and isort
