+++ Pipfile
@@ -4,13 +4,15 @@ url = "https://pypi.org/simple"
 verify_ssl = true
 
 [dev-packages]
-black = "==20.8b1"
+black = "==21.12b0"
+pytest = "*"
+isort = "*"
 
 [packages]
 requests = "*"
-requests-html = "*"
 semver = "*"
 docker = "*"
+beautifulsoup4 = "*"
 
 [requires]
-python_version = "3.8"
+python_version = "3.9"
