+++ build_versions/settings.py
@@ -1,0 +1,6 @@
+from pathlib import Path
+
+DOCKER_IMAGE_NAME = "nikolaik/python-nodejs"
+VERSIONS_PATH = Path("versions.json")
+DEFAULT_DISTRO = "buster"
+DISTROS = ["buster", "bullseye", "slim", "alpine"]
