+++ pyproject.toml
@@ -1,3 +1,8 @@
 [tool.black]
 line-length = 120
-target-version = ['py38']
+target-version = ['py39']
+
+[tool.isort]
+profile = "black"
+line_length = 120
+skip_gitignore = true
